+++ source/Form.js
@@ -5,7 +5,7 @@ import classNames from 'classnames'
 import Button from './Button'
 import { ModalContext } from './Modal'
 
-// Prevents `<form/> submission when `busy` is `true`.
+// Prevents `<form/> submission when `wait` is `true`.
 // And also inserts `<Form.Error/>` when `error` is passed.
 // Using `Component` here instead of `PureComponent`
 // because `<Form.Actions>` depends on `context` and therefore
@@ -33,6 +33,11 @@ export default class Form extends Component
 		post        : PropTypes.string,
 
 		// Is form submission in progress
+		wait        : PropTypes.bool.isRequired,
+
+		// (deprecated)
+		// (use `wait` instead)
+		// Is form submission in progress
 		busy        : PropTypes.bool.isRequired,
 
 		// CSS class
@@ -44,6 +49,7 @@ export default class Form extends Component
 
 	static defaultProps =
 	{
+		wait : false,
 		busy : false
 	}
 
@@ -117,12 +123,12 @@ export default class Form extends Component
 			event.preventDefault()
 		}
 
-		const { busy, submit } = this.props
+		const { wait, busy, submit } = this.props
 
 		// Prevent form double submit
 		// (because not only buttons submit a form,
 		//  therefore just disabling buttons is not enough).
-		if (busy)
+		if (wait || busy)
 		{
 			return
 		}